AP-3267 Move FullEmploymentDetailsController to Providers::Means namespace
