+++ app/services/flow/flows/provider_property.rb
@@ -21,12 +21,12 @@ module Flow
         shared_ownerships: {
           path: ->(application) { urls.providers_legal_aid_application_shared_ownership_path(application) },
           forward: lambda do |application|
-            if application.shared_ownership?
-              :percentage_homes
-            else
-              application.checking_answers? ? :restrictions : :vehicles
-            end
-          end,
+                     if application.shared_ownership?
+                       :percentage_homes
+                     else
+                       application.checking_answers? ? :restrictions : :vehicles
+                     end
+                   end,
           carry_on_sub_flow: true,
         },
         percentage_homes: {
