+++ app/views/providers/means/full_employment_details/show.html.erb
@@ -1,5 +1,5 @@
 <%= form_with(model: @form,
-              url: providers_legal_aid_application_full_employment_details_path(@legal_aid_application),
+              url: providers_legal_aid_application_means_full_employment_details_path(@legal_aid_application),
               method: :patch,
               local: true) do |form| %>
   <% page_heading = @legal_aid_application.has_multiple_employments? ? t('.page_heading_multi_jobs') : t('.page_heading_no_info') %>
