+++ app/views/shared/check_answers/_full_employment_details.html.erb
@@ -5,7 +5,7 @@
   <div class="govuk-grid-column-one-third govuk-summary-list--no-border align-text-right">
     <% unless read_only %>
       <p>
-        <%= link_to_accessible(t('generic.change'), providers_legal_aid_application_full_employment_details_path(@legal_aid_application),
+        <%= link_to_accessible(t('generic.change'), providers_legal_aid_application_means_full_employment_details_path(@legal_aid_application),
                                class: 'govuk-link change-link') %>
       </p>
     <% end %>
