+++ config/locales/en/providers.yml
@@ -633,6 +633,20 @@ en:
         form_of_service_type: Full representation
         substantive_certificate: Substantive certificate
         emergency_certificate: Emergency certificate
+    means:
+      full_employment_details:
+        show:
+          page_heading_no_info: HMRC has no record of your client's employment in the last 3 months
+          page_heading_multi_jobs: HMRC found a record of your client's employment
+          no_employment_info: You told us your client is employed but HMRC could not confirm this.
+          mutiple_employment_info: HMRC says your client had more than one job in the last 3 months.
+          enter_details: "Enter details of your client's employment in the last 3 months, including:"
+          details_1: how much they earned before tax
+          details_2: how much National Insurance they paid
+          details_3: how much tax they paid
+          details_4: any benefits in kind they received
+          employment_details_header: Enter your client's employment details
+          employment_details_hint: You'll need to upload supporting evidence later.
     means_summaries:
       show:
         h1-heading: Check your answers
@@ -744,20 +758,6 @@ en:
         debits-section-heading: Outgoings
         deductions-section-heading: Deductions
         proceeding_eligibility: Proceeding eligibility
-    full_employment_details:
-      show:
-        page_heading_no_info: HMRC has no record of your client's employment in the last 3 months
-        page_heading_multi_jobs: HMRC found a record of your client's employment
-        no_employment_info: You told us your client is employed but HMRC could not confirm this.
-        mutiple_employment_info: HMRC says your client had more than one job in the last 3 months.
-        enter_details: "Enter details of your client's employment in the last 3 months, including:"
-        details_1: how much they earned before tax
-        details_2: how much National Insurance they paid
-        details_3: how much tax they paid
-        details_4: any benefits in kind they received
-        employment_details_header: Enter your client's employment details
-        employment_details_hint: You'll need to upload supporting evidence later.
-
     no_income_summaries:
       show:
         page_heading: "Your client's income"
