+++ spec/requests/providers/client_completed_means_spec.rb
@@ -112,7 +112,7 @@ RSpec.describe Providers::ClientCompletedMeansController, type: :request do
 
             it "redirects to the no employed income page" do
               subject
-              expect(response).to redirect_to(providers_legal_aid_application_full_employment_details_path(legal_aid_application))
+              expect(response).to redirect_to(providers_legal_aid_application_means_full_employment_details_path(legal_aid_application))
             end
           end
 
@@ -124,7 +124,7 @@ RSpec.describe Providers::ClientCompletedMeansController, type: :request do
 
             it "redirects to the no employed income page" do
               subject
-              expect(response).to redirect_to(providers_legal_aid_application_full_employment_details_path(legal_aid_application))
+              expect(response).to redirect_to(providers_legal_aid_application_means_full_employment_details_path(legal_aid_application))
             end
           end
